fix(compat): fix StrictMode remount regression, split oversized test file
- next-script.tsx: the "already injected" guard was set at schedule
  time (before runWhenIdle) rather than at actual-injection time. React
  StrictMode's dev-only setup->cleanup->setup cycle cancels the pending
  deferred inject on the synthetic cleanup, then re-runs the effect —
  with the guard set early, that remount saw it already tripped and
  never rescheduled, so the script silently never injected under
  StrictMode for any deferred strategy (afterInteractive/lazyOnload/
  unset). Fixed by moving the flag into inject() itself, only set once
  it actually runs. (CodeRabbit review)
- Removed a redundant inner "already injected" check inside inject()
  itself — tracing through runWhenIdle's cancellation logic, inject()
  can never be invoked twice given the cancelled-flag guard already in
  place, so the extra check was unreachable defensive code.
- Split next-script.integration.test.tsx (350 non-comment lines) into
  three focused files per the project's 300-line SRP convention
  (.coderabbit.yaml): timing (strategy/idle/cancellation), rendering
  (src/children/attributes output), and strictmode (double-invoke
  correctness). Added the StrictMode + deferred-strategy regression
  test the prior single-file suite was missing — this is exactly the
  case that would have caught the bug above. 100% coverage maintained
  on next-script.tsx (27 tests total, up from 26).
